requirements.txt
streamlit>=1.28.0
feedparser>=6.0.10
google-generativeai>=0.3.0
PyGithub>=1.59.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0

